Fail when trying to load frr only features in native mode
Ideally, the checks of FRR only fields should happen in the FRR handler
of the speaker. Given the defaulting in place this is difficult to
achieve, and because it makes sense to have the defaulting collected in one
single place we inject another implementation driven hook to check the
raw config before any defaulting happens.

Current checks are meant to drop:
- ipv6 pools in case of native bgp mode
- ipv6 aggregation length
- BFD profiles

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>