FRR: Support eBGP multi-hop peering
Adding support for eBGP multi-hop peering in FRR is a must
as it is already supported by the native mode.

This is done by adding `ebgp-multihop` in FRR's config to a peer
that requires eBGP and is multi-hops away.

The problem with adding this without knowing if the peer is actually
multi-hops away breaks BFD, because BFD considers it multi-hop when
that parameter is set regardless if it actually is, and rejects the
peer in case it is not multi-hops away.

To solve that we introduce an auto-discovery mechanism to decide
if a given peer is multi-hops away or not, which is also overridable
by specifying the new multi-hop config parameter for a peer, and
setting `ebgp-multihop` accordingly.

Signed-off-by: Ori Braunshtein <obraunsh@redhat.com>