E2E: Change external containers setup to cover multi-hop eBGP
In order to cover multi-hop eBGP peering scenarios we need
to create an external eBGP container that is in a different
containers network.
To achieve that we create another container network (multi-hop-net)
as part of the suite for the eBGP container. To have connectivity
between the current containersNetwork and multi-hop-net we create
a new container that is in both, and route requests between the
networks through it by patching both the routing table of the speaker
nodes (speaker is host-networked) and the eBGP container.

The result is that the tests now run using 3 external FRR peers:
* 2 that are a single-hop away from the speaker - iBGP and eBGP.
* 1 multi-hops away from the speaker - eBGP.

Signed-off-by: Ori Braunshtein <obraunsh@redhat.com>