Replace the custom controller logic with controller-runtime
Here we do the wiring between the added controllers and the k8s
independent logic scattered across the internal folder.
We preserve the handlers, and we change the way we feed them.

A new set of controllers listen for config, nodes, services, and call
the previous callbacks.

No need to wait for the controllers to be synced, because now this is
handled by the manager.

The corner case where we ask for reprocessing all the services is
handled by the ServiceReload controller, which listens to a channel for
events. It's one way to avoid chainging the existing logic, because it
assumes one controller in charge of the reload and other controllers
able to signal it.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>